# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Base Restful API Documentation",
        "description": "Restful API documentation",
        "contact": {
            "email": "qrizan@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Base Restful API Documentation"
        }
    ],
    "paths": {
...
        "/api/admin/permissons": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get permissons",
                "operationId": "ceeabe29de1324295bf7f3f281ef2c2c",
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/admin/permissons/all": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all permissons",
                "operationId": "20690d5033cf7ca7be86bf703aba05ee",
                "responses": {
                    "201": {
                        "description": "Success"
                    },
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}